Исправлены ошибки в подсистеме лицензирования. Добавлены контролллеры для CRUD операций для каждой сущности. Была пройдена полная проверка всей системы.

+++ src/main/java/com/mtuci/lazarev/models/Ticket.java
@@ -1,12 +1,13 @@
 package com.mtuci.lazarev.models;
 
 import lombok.Data;
+
 import java.sql.Date;
 
 @Data
 public class Ticket {
     private Date nowDate, activationDate, expirationDate;
-    private Integer expiration;
+    private Long expiration;
 
     private Long userID, deviceID;
     private boolean isBlockedLicence;
